requirements-dev.txt
Lin-CMS==0.4.11
pillow==9.3.0
flask-cors==3.0.10
gunicorn==20.1.0
gevent==21.8.0
flask-socketio==5.1.1
blinker==1.4
python-dotenv==0.19.2
gevent-websocket==0.10.1
flask-sqlacodegen==1.1.8
black==22.3.0
isort==5.10.1
watchdog==2.1.6
coverage==6.3.2
pytest==6.2.5
pre-commit==2.17.0
pytest-ordering==0.6
flake8==4.0.1
mypy==0.950
Werkzeug==2.1.2

requests
pymysql
tqdm
# PyExecJS
# psutil==5.9.8
